+++ server/src/app/app.module.ts
@@ -1,12 +1,12 @@
 import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { PrismaModule } from './prisma/prisma.module';
-import { UserModule } from './user/user.module';
-import { AuthModule } from './auth/auth.module';
-import { HealthModule } from './health/health.module';
-import ERApiController from './erapi.controller';
-import { EldenRingModule } from './eldenring.module';
+import { PrismaModule } from '../prisma/prisma.module';
+import { UserModule } from '../user/user.module';
+import { AuthModule } from '../auth/auth.module';
+import { HealthModule } from '../health/health.module';
+import ERApiController from '../erapi/erapi.controller';
+import { EldenRingModule } from '../erapi/erapi.module';
 
 @Module({
   imports: [
