+++ server/src/main.ts
@@ -5,7 +5,7 @@ import {
   DocumentBuilder,
   SwaggerCustomOptions,
 } from '@nestjs/swagger';
-import { AppModule } from './app.module';
+import { AppModule } from './app/app.module';
 import { PrismaService } from './prisma/prisma.service';
 import * as expressSession from 'express-session';
 import { PrismaSessionStore } from '@quixo3/prisma-session-store';
@@ -14,7 +14,7 @@ import * as passport from 'passport';
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule, {
-    logger: ['error', 'warn', 'log'],
+    logger: ['error', 'warn', 'debug'],
   });
   const prismaService = app.get(PrismaService);
   await prismaService.enableShutdownHooks(app);
