--- server/utils/logging.ts
@@ -1,52 +1,0 @@
-/* eslint-disable no-console */
-const getTimeStamp = (): string => {
-  return new Date().toISOString();
-};
-
-const info = (namespace: string, message: string, object?: unknown) => {
-  if (object) {
-    console.log(`[${getTimeStamp()}] [INFO] [${namespace}] ${message}`, object);
-  } else {
-    console.log(`[${getTimeStamp()}] [INFO] [${namespace}] ${message}`);
-  }
-};
-
-const warn = (namespace: string, message: string, object?: unknown) => {
-  if (object) {
-    console.warn(
-      `[${getTimeStamp()}] [WARN] [${namespace}] ${message}`,
-      object,
-    );
-  } else {
-    console.warn(`[${getTimeStamp()}] [WARN] [${namespace}] ${message}`);
-  }
-};
-
-const debug = (namespace: string, message: string, object?: unknown) => {
-  if (object) {
-    console.debug(
-      `[${getTimeStamp()}] [DEBUG] [${namespace}] ${message}`,
-      object,
-    );
-  } else {
-    console.debug(`[${getTimeStamp()}] [DEBUG] [${namespace}] ${message}`);
-  }
-};
-
-const error = (namespace: string, message: string, object?: unknown) => {
-  if (object) {
-    console.error(
-      `[${getTimeStamp()}] [ERROR] [${namespace}] ${message}`,
-      object,
-    );
-  } else {
-    console.error(`[${getTimeStamp()}] [ERROR] [${namespace}] ${message}`);
-  }
-};
-
-export default {
-  info,
-  warn,
-  error,
-  debug,
-};
